misc: docs + small refactoring
