+++ grc20-core/src/mapping/query_utils/mod.rs
@@ -6,14 +6,12 @@ pub mod order_by;
 pub mod prop_filter;
 pub mod query_builder;
 pub mod query_part;
-pub mod types_filter;
 pub mod version_filter;
 
 pub use attributes_filter::AttributeFilter;
 pub use order_by::{FieldOrderBy, OrderDirection};
 pub use prop_filter::PropFilter;
 pub use query_part::QueryPart;
-pub use types_filter::TypesFilter;
 pub use version_filter::VersionFilter;
 
 pub trait Query<T>: Sized {
